+++ rapdev_nutanix/datadog_checks/rapdev_nutanix/__about__.py
@@ -1,3 +1,3 @@
 # (C) RapDev, Inc. 2020-present
 # All rights reserved
-__version__ = '1.3.5'
+__version__ = '1.3.6'